+++ src/typing_ex/typing.cljs
@@ -6,38 +6,41 @@
    [cljs.reader :refer [read-string]]
    [cljs.core.async :refer [<!]]
    [clojure.string :as str]
-   [reagent.core :refer [atom]]
+   [reagent.core :as r]
    [reagent.dom :as rdom]
+   [taoensso.timbre :as timbre]
    [typing-ex.plot :refer [plot]]))
 
-(def ^:private version "1.5.5")
+(def ^:private version "1.7.0")
 
-(defonce app-state (atom {:text "wait a little"
-                          :answer ""
-                          :seconds 60
-                          :errors 0}))
+(def ^:private timeout 60)
+(def ^:private todays-max 10)
 
-(defonce first-key (atom false))
-
-(defonce todays (atom {}))
+(defonce todays-trials (r/atom 0))
+(defonce app-state (r/atom {}))
+(defonce todays (r/atom {}))
 
 (defn get-login []
   (-> (.getElementById js/document "login")
       (.-value)))
 
 (defn reset-app-state! []
-  (go (let [response (<! (http/get (str "/drill")))
-            {s :body} (<! (http/get (str "/todays/" (get-login))))]
-        (swap! app-state assoc :text (:body response)
-               :answer ""
-               :seconds 60
-               :errors 0)
-        ;;(.log js/console "text" s)
-        (reset! first-key false)
-        (reset! todays (->> (read-string s)
-                            (map #(assoc % :pt (max 0 (:pt %)))))))))
+  (go (let [{drill :body}  (<! (http/get (str "/drill")))
+            words (str/split drill #"\s+")
+            {scores :body} (<! (http/get (str "/todays/" (get-login))))]
+        (reset! todays (read-string scores))
+        (reset! app-state
+                {:text drill
+                 :answer ""
+                 :seconds timeout
+                 :errors 0
+                 :words words
+                 :words-max (count words)
+                 :pos 0
+                 :results []})
+        (.focus (.getElementById js/document "drill")))))
 
-;;; pt will not be nagative.
+;;; pt must not be nagative.
 (defn pt-raw [{:keys [text answer seconds errors]}]
   (let [s1 (str/split text #"\s+")
         s2 (str/split answer #"\s+")
@@ -47,7 +50,8 @@
         bads  (count (remove (fn [[x y]] (= x y)) s1<>s2))
         err   (* -1 errors errors)
         score (int (* 100 (- (/ goods all) (/ bads goods))))]
-    (js/console.log "goods bads all error score: " goods bads all err score)
+    ;;(js/console.log "goods bads all error score: " goods bads all err score)
+    (timbre/info (get-login) goods bads all err score)
     (if (= all (+ goods bads))
       (+ score err seconds)
       (+ score err))))
@@ -55,10 +59,7 @@
 (defn pt [args]
   (max 0 (pt-raw args)))
 
-(defonce todays-count (atom 0))
-(def ^:private todays-max 10)
-
-(defn login-pt-message [{:keys [pt login]}]
+(defn intermission-message [{:keys [pt login]}]
   (let [s1 (str login " さんのスコアは " pt " 点です。")
         s2 (condp <= pt
              100 "すばらしい。最高点取れた？平均で 80 点越えよう。"
@@ -66,54 +67,66 @@
              60 "だいぶ上手です。この調子でがんばれ。"
              30 "指先を見ずに、ゆっくり、ミスを少なく。"
              "練習あるのみ。")]
-    (swap! todays-count inc)
-    (if (< todays-max @todays-count)
-      (do
-        (reset! todays-count 0)
-        (str s1 "\n" s2 "\n" "いったん休憩入れようか？"))
-      (str s1 "\n" s2))))
+    (js/alert s1 "\n" s2)))
 
-(defn send-score []
-  (go (let [token (.-value (js/document.getElementById "__anti-forgery-token"))
-            response (<! (http/post
-                          "/score"
-                          {:form-params
-                           {:pt (pt @app-state)
-                            :__anti-forgery-token token}}))]
-        (reset-app-state!)
-        (js/alert (login-pt-message (read-string (:body response))))
-        (.focus (.getElementById js/document "drill")))))
+(defn send-score! []
+  (go (let [token (-> (js/document.getElementById "__anti-forgery-token")
+                      .-value)
+            {body :body} (<! (http/post
+                              "/score"
+                              {:form-params
+                               {:pt (pt @app-state)
+                                :__anti-forgery-token token}}))]
+        (intermission-message (read-string body))
+        (swap! todays-trials inc)
+        (when (zero? (mod @todays-trials todays-max))
+          (js/alert "いったん休憩入れよう 🍵")))))
 
-(defn count-down []
-  (when true ;; @first-key
-    (swap! app-state update :seconds dec))
+(defn countdown []
+  (swap! app-state update :seconds dec)
   (when (zero? (:seconds @app-state))
-    (send-score)))
+    (if (zero? (count (:answer @app-state)))
+      (js/alert "タイプ忘れた？")
+      (send-score!))
+    (reset-app-state!)))
 
 ;; FIXME: when moving below block to top of this code,
-;;        becomes not counting down.
-;;(declare count-down)
-(defonce updater (js/setInterval count-down 1000))
+;;        becomes not counting down even if declared.
+;;(declare countdown)
+(defonce updater (js/setInterval countdown 1000))
 
-;; FIXME: function name
-(defn show-sorry [n]
-  (take n (repeat "🙅"))) ;;🙅💧💦💔❌🦠🥶🥺
+(defn check-word []
+  (let [target (get (@app-state :words) (@app-state :pos))
+        typed  (last (str/split (@app-state :answer) #"\s+"))]
+    (.log js/console target typed)
+    (swap! app-state update :results
+           #(conj % (if (= target typed) "🟢" "🔴")))
+    (swap! app-state update :pos inc)
+    (when (<= (@app-state :words-max) (@app-state :pos))
+      (send-score!)
+      (reset-app-state!))))
 
 (defn check-key [key]
-  ;; (when-not @first-key
-  ;;   (swap! first-key not))
-  (when (= key "Backspace")
-    (swap! app-state update :errors inc)))
+  (case key
+    " " (check-word)
+    "Enter" (check-word)
+    "Backspace" (swap! app-state update :errors inc)
+    nil))
 
+;;🙅💧💦💔❌🦠🥶🥺
 (defn error-component []
-  [:p "　" (show-sorry (:errors @app-state))])
+  (.log js/console "errors" (:errors @app-state))
+  [:div.drill (repeat (:errors @app-state) "💔")])
+
+(defn results-component []
+  [:div.drill (apply str (@app-state :results))])
 
 (defn ex-page []
   [:div
    [:h2 "Typing: Challenge"]
    [:p
     {:class "red"}
-    "指先見ないで、ゆっくり、確実に。単語間のスペースは一個でよい。"]
+    "指先見ないで、ゆっくり、確実に。単語間のスペースは一個で。"]
    [:pre {:id "example"} (:text @app-state)]
    [:textarea {:name "answer"
                :id "drill"
@@ -124,13 +137,15 @@
                                   :answer
                                   (-> % .-target .-value))}]
    [error-component]
+   [results-component]
    [:p
     [:input {:type  "button"
              :id    "seconds"
              :class "btn btn-success btn-sm"
              :style {:font-family "monospace"}
              :value (:seconds @app-state)
-             :on-click send-score}] " 🔚全部打ち終わってクリックするとボーナス"]
+             :on-click #(do (send-score!) (reset-app-state!))}]
+    " 🔚クリックしなくても全部打った後にスペースかエンターでボーナス"]
    [:p
     "Your todays:"
     [:br]
@@ -144,11 +159,11 @@
    [:div "hkimura, " version]])
 
 (defn start []
+  (reset-app-state!)
   (rdom/render [ex-page] (js/document.getElementById "app"))
   (.focus (.getElementById js/document "drill")))
 
 (defn ^:export init []
-  (reset-app-state!)
   ;; init is called ONCE when the page loads
   ;; this is called in the index.html and must be exported
   ;; so it is available even in :advanced release builds